pyproject.toml
[project]
name = "matvey-portfolio-bot"
version = "0.1.0"
description = "Telegram portfolio bot for Matvey Bliznyuk - Python developer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "flake8>=7.0.0",
    "flake8-pyproject>=1.2.0",
    "mypy>=1.0.0",
    "black>=24.0.0",
]

[tool.flake8]
max-line-length = 120
